refactor(root): move jsx renderer to src/jsx/ folder
Move jsx-render.ts to src/jsx/ and add a jsx.ts barrel export so the
tsup entry for @blinkk/root/jsx is defined from src/jsx/jsx.ts.
