+++ packages/root/src/jsx/jsx.ts
@@ -1,0 +1,2 @@
+export {renderJsxToString} from './jsx-render.js';
+export type {JsxRenderOptions} from './jsx-render.js';
