+++ packages/root/tsup.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
     cli: 'src/cli/cli.ts',
     core: 'src/core/core.ts',
     functions: 'src/functions/functions.ts',
-    jsx: 'src/render/jsx-render.ts',
+    jsx: 'src/jsx/jsx.ts',
     middleware: 'src/middleware/middleware.ts',
     node: 'src/node/node.ts',
     render: 'src/render/render.tsx',
